+++ src/lib/kotlin/slatekit-meta/src/main/kotlin/slatekit/meta/models/Model.kt
@@ -13,17 +13,13 @@
 
 package slatekit.meta.models
 
-import slatekit.common.DateTime
+import slatekit.common.Field
+import slatekit.common.Id
 import slatekit.common.naming.Namer
-import slatekit.meta.KTypes
-//import java.time.*
-import org.threeten.bp.*
-import slatekit.common.DateTimes
 import slatekit.common.ext.orElse
+import slatekit.meta.Reflector
 import slatekit.meta.Schema
-import java.util.*
 import kotlin.reflect.KClass
-import kotlin.reflect.KProperty
 
 /**
  * Stores the schema of a data-model with properties.
@@ -94,6 +90,42 @@ class Model(
             val schema = Schema<TId, T>(idType, tType)
             builder(schema)
             return schema.model
+        }
+
+        /**
+         * Builds a schema ( Model ) from the Class/Type supplied.
+         * NOTE: The mapper then works off the Model class for to/from mapping of data to model.
+         * @param dataType
+         * @return
+         */
+        @JvmStatic
+        fun loadSchema(dataType: KClass<*>, idFieldName: String? = null, namer: Namer? = null, table: String? = null): Model {
+            val modelName = dataType.simpleName ?: ""
+            val modelNameFull = dataType.qualifiedName ?: ""
+
+            // Get Id
+            val idFields = Reflector.getAnnotatedProps<Id>(dataType, Id::class)
+            val idField = idFields.firstOrNull()
+
+            // Now add all the fields.
+            val matchedFields = Reflector.getAnnotatedProps<Field>(dataType, Field::class)
+
+            // Loop through each field
+            val withAnnos = matchedFields.filter { it.second != null }
+            val fields = withAnnos.map { matchedField ->
+                val modelField = ModelField.ofData(matchedField.first, matchedField.second!!,
+                        namer, idField == null, idFieldName)
+                val finalModelField = if (!modelField.isBasicType()) {
+                    val model = loadSchema(modelField.dataCls, namer = namer)
+                    modelField.copy(model = model)
+                } else modelField
+                finalModelField
+            }
+            val allFields = when(idField) {
+                null -> fields
+                else -> mutableListOf(ModelField.ofId(idField.first, "", namer)).plus(fields)
+            }
+            return Model(modelName, modelNameFull, dataType, modelFields = allFields, namer = namer, tableName = table ?: modelName)
         }
     }
 }
