--- src/lib/kotlin/slatekit-meta/src/main/kotlin/slatekit/meta/models/ModelMapperSettings.kt
@@ -1,3 +1,0 @@
-package slatekit.meta.models
-
-class ModelMapperSettings(val persistUTCDate: Boolean = true)
